src/components/Spinner/index.tsx
/** @jsxImportSource @emotion/react */
import { keyframes } from "@emotion/react";
import styled from "@emotion/styled";

const spin = keyframes`
  0% { transform: rotate(0deg); }
  100% { transform: rotate(360deg); }
`;

const SpinnerContainer = styled.div`
  display: flex;
  align-items: center;
  justify-content: center;
  height: 50px;
`;

const Spinner = styled.div`
  border: 4px solid #f3f3f3;
  border-top: 4px solid #334155;
  border-radius: 50%;
  width: 24px;
  height: 24px;
  animation: ${spin} 1s linear infinite;
`;

const LoadingAnimation = () => {
  return (
    <SpinnerContainer data-testid="spinner-container">
      <Spinner data-testid="spinner" />
    </SpinnerContainer>
  );
};

export default LoadingAnimation;
